cross entropy working, but still got to do code cleanup

+++ dlgrad/src/c/index.c
@@ -3,7 +3,7 @@
 
 
 // only for nll loss
-void indexing(float *x, float *out, int *xshape, int *xstride, int *idxs)
+void indexing(float *x, float *out, int *xshape, int *xstride, float *idxs)
 {
     for (int i=0; i<xshape[0]; i++) {   // rows
         int out_idx = idxs[i]*xstride[1] + i*xstride[0];
