+++ dlgrad/src/c/index.h
@@ -2,6 +2,6 @@
 #define INDEX
 
 
-void indexing(float *x, float *out, int *xshape, int *xstride, int *idxs);
+void indexing(float *x, float *out, int *xshape, int *xstride, float *idxs);
 
 #endif
